Generate the stub app bundle
This builds and copies the stub binary into the Mac app resources when building. Then before installing the stub app, it creates a minimal app bundle with that binary and codesigns it. As far as what JungleGym does, it should be possible to sandbox the app, but doing so breaks FBSimulatorControl. It looks like it can still create and boot a simulator (from the logs at least) but it gets hung up installing or running the app.