Build the stub with SPM
Because the stub binary is copied into a new app package, it needs to be statically linked with the Swift standard library. It doesn't seem like Xcode will let you do this, but SPM does. It took some experimenting to get it to link against the simulator SDK frameworks. This feels kinda bleeding edge (read: will be painful for future Brandon) but it's neat that this is possible. The simpler, safer alternative would be to build the stub with Objective-C, and then we could keep the Xcode project and the dev process when making changes to it would be nicer.